bedocat/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
mysqlclient==2.2.4
pillow==10.3.0
Serializer==0.2.1
simplejson==3.19.2
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.6.0
